groq_ai/requirements.txt
langchain-groq
python-dotenv
requests
chromadb
langchain-community
beautifulsoup4
pandas
streamlit

